+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release Notes for Algolia
 
+## 3.0.0 - 2021-06-29
+
+### Added
+* Helper to convert an object of filters into the stringified Algolia syntax for filtered searches
+
+### Changed
+* Algolia PHP SDK updated to 3.x
+* PHP 7.1+ required
+* Craft 3.1.19+ required
+
 ## 2.0.0 - 2019-03-15
 
 Initial Craft 3 release